[Runtime] Move some of the changes back to int.
Some of the fixes done for the warnings have breaking changes. Move back
to int and ensure that we do not have any compilation errors (we are
using -Werror).

Fixes: https://github.com/xamarin/xamarin-macios/issues/7509

+++ runtime/trampolines.m
@@ -443,7 +443,7 @@ xamarin_collapse_struct_name (const char *type, char struct_name[], int max_char
 	return true;
 }
 
-unsigned long
+int 
 xamarin_get_frame_length (id self, SEL sel)
 {
 	if (self == NULL)
@@ -510,7 +510,7 @@ xamarin_get_frame_length (id self, SEL sel)
 	// we can't detect varargs, so just add 16 more pointer sized arguments to be on the safe-ish side.
 	length += sizeof (void *) * 16;
 
-	return length;
+	return (int) length;
 }
 
 static inline void
